# expression: implement vectorized evaluation for `builtinJSONMergeSig`