<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->
### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for `builtinJSONMergeSig` mentioned in https://github.com/pingcap/tidb/issues/12104

### What is changed and how it works?
about 40% faster
```
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONMergeSig-VecBuiltinFunc-4                      482           2282113 ns/op          949025 B/op      14082 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONMergeSig-NonVecBuiltinFunc-4                   386           3197396 ns/op         1019904 B/op      14752 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test