@js00070 2019-10-28
@XuHuaiyu 
I did several times of benchstat, the results are unstable because of my CPU condition(the frequency is changing on different temperature), but we can see my original code is faster than the approch like you mentioned.

```
name                                                                old time/op    new time/op    delta
VectorizedBuiltinJSONFunc/builtinJSONMergeSig-VecBuiltinFunc-12        822µs ± 1%    1064µs ± 0%  +29.47%  (p=0.000 n=19+17)
VectorizedBuiltinJSONFunc/builtinJSONMergeSig-NonVecBuiltinFunc-12    1.08ms ± 1%    1.30ms ± 1%  +20.19%  (p=0.000 n=19+20)
```
```
name                                                                old time/op    new time/op    delta
VectorizedBuiltinJSONFunc/builtinJSONMergeSig-VecBuiltinFunc-12        906µs ± 1%    1064µs ± 0%  +17.38%  (p=0.000 n=19+17)
VectorizedBuiltinJSONFunc/builtinJSONMergeSig-NonVecBuiltinFunc-12    1.18ms ± 1%    1.30ms ± 1%   +9.60%  (p=0.000 n=19+20)
```
```
name                                                                old time/op    new time/op    delta
VectorizedBuiltinJSONFunc/builtinJSONMergeSig-VecBuiltinFunc-12       1.02ms ± 1%    1.06ms ± 0%   +4.59%  (p=0.000 n=19+17)
VectorizedBuiltinJSONFunc/builtinJSONMergeSig-NonVecBuiltinFunc-12    1.32ms ± 1%    1.30ms ± 1%   -1.87%  (p=0.000 n=19+20)
```

the code which produces `new.txt` is as follow
```go
	result.ReserveJSON(nr)
	for i := 0; i < nr; i++ {
		nullFlag := false
		jsonVal := make([]json.BinaryJSON, 0, len(b.args))
		for j := 0; j < len(b.args); j++ {
			if argBuffers[j].IsNull(i) {
				result.AppendNull()
				nullFlag = true
				break
			}
			jsonVal = append(jsonVal, argBuffers[j].GetJSON(i))
		}
		if !nullFlag {
			result.AppendJSON(json.MergeBinary(jsonVal))
		}
	}
```